requirements.txt
fastapi
uvicorn
openai
requests
pydantic
python-dotenv
openenv-core
